requirements.txt
fastapi
uvicorn
jinja2
python-multipart
PyMuPDF
python-docx
spacy
pytube
